# swagger/swagger.yml
swagger: "2.0"
info:
  description: Projeto para <strong>Processo Seletivo Engenheiro(a) de Dados</strong><br/><br/>
    Esta interface serve para auxiliar na navegação das rotas.<br/>
    - [<strong>Carregar Dados</strong>] passo inicial para realização da carga das informações fornecidas;<br/>
    - [<strong>Avaliação de concorrentes</strong>] representação das informações sobre fluxos de pessoas;<br/>
    - [<strong>Avaliação por seguimento</strong>] representação comparativa das empresas por seguimento de atuação;<br/>
    - [<strong>Bairros</strong>] representa cadastro de municipios;<br/>
    - [<strong>Concorrentes</strong>] representa cadastro de empresas;<br/>
    - [<strong>Eventos</strong>] representa os logs coletados dos fluxos de pessoas;<br/>
    - [<strong>Precopraticado</strong>] representa uma rota para cadastro de preços dos concorrentes;<br/>    

  version: "1.0.0"
  title: Avaliação de Concorrentes
consumes:
  - application/json
produces:
  - application/json

#basePath: /

# Paths suportados pela aplicação
paths:
...
  /precopraticado/v1:
    get:
      operationId: api.precopraticados.get_precopraticado
      tags:
        - precopraticados
      summary: Recuperando todos preços praticados
      description: Mostrando preços praticados
      responses:
        200:
          description: Visualizando todos preços praticados
    post:
      operationId: api.precopraticados.add_precopraticado
      tags:
        - precopraticados
      summary: Incluindo preço praticado
      description: Incluindo um único registro
      parameters:
        - name: precopraticado
          in: body
          description: criando preço praticado
          required: true
          schema:
            type: object
            properties:
              faixa_preco:
                type: integer
              descricao:
                type: string
      responses:
        200:
          description: Sucesso na criação preço praticado
        400:
          description: Invalida requisição
  /precopraticado/v1/{faixa_preco}:
    get:
      operationId: api.precopraticados.get_by_faixa_preco
      tags:
        - precopraticados
      summary: Recuperando preço praticado pela faixa de preço
      description: Visualização por faixa de preço
      parameters:
        - name: faixa_preco
          in: path
          description: recuperando dados
          type: integer
          required: true
      responses:
        200:
          description: Sucesso na visualização por faixa de preço
          schema:
            type: array
            items:
              properties:
                faixa_preco:
                  type: integer
                descricao:
                  type: string
        404:
          description: Faixa de preço não encontrado

    put:
      operationId: api.precopraticados.update_descricao
      tags:
        - precopraticados
      summary: Atualizar faixa de preço
      description: Atualizar uma unica faixa de preço
      parameters:
        - name: faixa_preco
          in: path
          description: faixa de preço para atualizar
          type: integer
          required: true
        - name: descricao
          in: body
          description: campo para atualizar
          required: true
          schema:
            type: object
            properties:
              descricao:
                type: string
      responses:
        204:
          description: Sucesso na atualização da descricao da faixa de preço
        400:
          description: Invalid request

    delete:
      operationId: api.precopraticados.delete_precopraticado
      tags:
        - precopraticados
      summary: Excluir preço praticado pela faixa de preço
      description: Excluir por faixa de preço
      parameters:
        - name: faixa_preco
          in: path
          description: Exclur faixa de preço
          type: integer
          required: true
      responses:
        204:
          description: Sucessfully preço praticado
        404:
          description: Preco praticado não encontrado, indisponivel para excluir.